requirements.txt
# Core Frameworks
streamlit==1.36.0
python-dotenv==1.0.1

# LangChain Libraries - Pinned for stability
langchain==0.2.0
langchain-community==0.2.0
langchain-core==0.2.1
langchain-groq==0.1.5
langchain-openai==0.1.7
langchain-google-genai==1.0.5
langchain-huggingface==0.0.3
langchainhub==0.1.18  # <-- CRITICAL FIX: For the ReAct agent

# RAG & Document Processing
faiss-cpu==1.8.0
sentence-transformers==2.7.0
pypdf==4.2.0
unstructured==0.14.4
nltk==3.8.1

# Tools & Utilities
tavily-python==0.3.3
gTTS==2.5.1
certifi==2024.7.4  # <-- CRITICAL FIX: For SSL errors

# Dependency Warning Fix
urllib3==2.2.1
charset-normalizer==3.3.2
requests==2.31.0